+++ services/QuillLMS/client/app/bundles/Staff/components/evidence/activitySessions/sessionsIndex.tsx
@@ -6,7 +6,8 @@ import * as moment from 'moment';
 import { firstBy } from 'thenby';
 import DateTimePicker from 'react-datetime-picker';
 
-import { handlePageFilterClick, renderHeader } from "../../../helpers/evidence";
+import { handlePageFilterClick } from "../../../helpers/evidence";
+import { renderHeader } from "../../../helpers/evidence/renderHelpers";
 import { Error, Spinner, DropdownInput, Input } from '../../../../Shared/index';
 import { fetchActivity, fetchActivitySessions } from '../../../utils/evidence/activityAPIs';
 import { DropdownObjectInterface, ActivitySessionInterface, ActivitySessionsInterface, InputEvent } from '../../../interfaces/evidenceInterfaces';
