+++ services/QuillLMS/client/app/bundles/Staff/components/evidence/configureRules/ruleForm.tsx
@@ -9,7 +9,7 @@ import RuleUniversalAttributes from './ruleUniversalAttributes';
 
 import { fetchRules, fetchUniversalRules } from '../../../utils/evidence/ruleAPIs';
 import { fetchConcepts, } from '../../../utils/evidence/conceptAPIs';
-import { renderErrorsContainer } from '../../../helpers/evidence';
+import { renderErrorsContainer } from '../../../helpers/evidence/renderHelpers';
 import { formatPrompts } from '../../../helpers/evidence/promptHelpers';
 import { handleSubmitRule, getInitialRuleType, formatInitialFeedbacks, returnInitialFeedback, formatRegexRules } from '../../../helpers/evidence/ruleHelpers';
 import { ruleOptimalOptions, regexRuleTypes, PLAGIARISM } from '../../../../../constants/evidence';
