+++ services/QuillLMS/client/app/bundles/Staff/components/evidence/configureRules/rulesIndexRouter.tsx
@@ -4,7 +4,7 @@ import { useQuery } from 'react-query';
 
 import Rules from './rules';
 
-import { renderHeader } from '../../../helpers/evidence';
+import { renderHeader } from '../../../helpers/evidence/renderHelpers';
 import { getPromptForComponent } from '../../../helpers/evidence/promptHelpers';
 import { BECAUSE, BUT, SO, ALL } from '../../../../../constants/evidence';
 import { fetchActivity } from '../../../utils/evidence/activityAPIs';
