+++ services/QuillLMS/client/app/bundles/Staff/components/evidence/regexRules/regexRulesIndex.tsx
@@ -3,7 +3,7 @@ import { Link, RouteComponentProps } from 'react-router-dom'
 import { queryCache, useQuery } from 'react-query';
 import { firstBy } from 'thenby';
 
-import { renderHeader } from '../../../helpers/evidence';
+import { renderHeader } from '../../../helpers/evidence/renderHelpers';
 import { getPromptsIcons } from '../../../helpers/evidence/promptHelpers';
 import { getPromptIdString } from '../../../helpers/evidence/ruleHelpers';
 import { ActivityRouteProps, RuleInterface, RegexRuleInterface } from '../../../interfaces/evidenceInterfaces';
