+++ services/QuillLMS/client/app/bundles/Staff/helpers/evidence.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import stripHtml from "string-strip-html";
 
 import {
@@ -14,18 +13,7 @@ import {
   PLAGIARISM,
   FLAG
 } from '../../../constants/evidence';
-import { ActivityInterface, DropdownObjectInterface } from '../interfaces/evidenceInterfaces'
-
-const quillCheckmark = `/images/green_check.svg`;
-const quillX = '/images/red_x.svg';
-
-export const getCheckIcon = (value: boolean) => {
-  if(value) {
-    return (<img alt="quill-circle-checkmark" src={quillCheckmark} />)
-  } else {
-    return (<img alt="quill-circle-checkmark" src={quillX} />);
-  }
-}
+import { DropdownObjectInterface } from '../interfaces/evidenceInterfaces'
 
 export const buildActivity = ({
   activityFlag,
@@ -182,44 +170,4 @@ export const validateForm = (keys: string[], state: any[], ruleType?: string) =>
 
 export function titleCase(string: string){
   return string[0].toUpperCase() + string.slice(1).toLowerCase();
-}
-
-export const renderIDorUID = (idOrRuleId: string | number, type: string) => {
-  return(
-    <section className="label-status-container">
-      <p id="label-status-label">{type}</p>
-      <p id="label-status">{idOrRuleId}</p>
-    </section>
-  );
-}
-
-export function renderErrorsContainer(formErrorsPresent: boolean, requestErrors: string[]) {
-  if(formErrorsPresent) {
-    return(
-      <div className="error-message-container">
-        <p className="all-errors-message">Please check that all fields have been completed correctly.</p>
-      </div>
-    );
-  }
-  return(
-    <div className="error-message-container">
-      {requestErrors.map((error, i) => {
-        return <p className="all-errors-message" key={i}>{error}</p>
-      })}
-    </div>
-  )
-}
-
-export const renderHeader = (activityData: {activity: ActivityInterface}, header: string) => {
-  if(!activityData) { return }
-  if(!activityData.activity) { return }
-  const { activity } = activityData;
-  const { title, notes } = activity;
-  return(
-    <section className="comprehension-page-header-container">
-      <h2>{header}</h2>
-      <h3>{title}</h3>
-      <h4>{notes}</h4>
-    </section>
-  );
 }
